rename: chagne path and img name (#81)
